Switch to pulseaudio's git-version-gen script
Note: scriptversion=2012-09-25.20

This will allow us to avoid the versioning
patterns with two or three dashes, which
conflict with Debian versioning schemes.

For example: debhelper, git-buildpackage,
and dch tools all like to assume that a
dash separates the Debian / Ubuntu part
of the version string from the Upstream
part. The Upstream part will be generated
by this script, from a git release tag.

Reference:

https://answers.launchpad.net/launchpad/+question/691015