Vendor autoconf-archive AX_IS_RELEASE & AX_COMPILER_FLAGS M4 templates
Check for the presence of a git repository to determine whether things
like compiler warnings should be fatal. The warnings themselves will be
hooked up later.

This adds an implicit dependency on autoconf-archive or GNOME’s
m4-common. These are pulled in by debuild, and the relevant macros are
automatically distributed in tarballs — so the dependencies only need to
be explicitly installed if you are building from git without using
debuild or pbuilder. To avoid creating a hard dependency on these,
and to avoid potential API breakage introduced by upstream changes,
these m4 macros have been vendored in place in this repo.

Note that AC_CONFIG_MACRO_DIR() has been added to configure.ac.