Add clean-local-am autoreconf clean target
This should clean all files that aclocal, autoconf, automake, & autoreconf
generate.  These are not tracked in the git repository, and depend on the
version of these tools.  To make maintaining this project easier, they
are not added to the SCM repo and you can use the clean-local-am target
to remove them.

Reference:

https://lists.gnu.org/archive/html/autoconf/2007-03/msg00043.html